requirements.txt
# Data Processing and Analysis
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2

# API and Web
fastapi==0.95.2
uvicorn==0.22.0
streamlit==1.22.0

# Database
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0

# Machine Learning
scikit-learn==1.2.2
transformers==4.30.2
torch==2.0.1

# Data Visualization
plotly>=5.19.0

# Environment and Testing
pytest>=8.0.0

# Utilities
jupyter>=1.0.0
notebook>=7.0.0
tqdm>=4.66.0 